+++ token-lending/oracles/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "oracles"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+pyth-sdk-solana = "0.8.0"
+pyth-solana-receiver-sdk = "0.3.0"
+solana-program = ">=1.9"
+anchor-lang = "0.28.0"
+solend-sdk = { path = "../sdk" }
+switchboard-on-demand = "0.1.12"
+switchboard-program = "0.2.0"
+switchboard-v2 = "0.1.3"
+
+[dev-dependencies]
+bytemuck = "1.5.1"
+assert_matches = "1.5.0"
+base64 = "0.13"
+log = "0.4.14"
+proptest = "1.0"
+solana-sdk = ">=1.9"
+serde = "1.0.140"
+serde_yaml = "0.8"
+rand = "0.8.5"
