+++ token-lending/program/Cargo.toml
@@ -13,15 +13,13 @@ test-bpf = []
 
 [dependencies]
 bytemuck = "1.5.1"
-pyth-sdk-solana = "0.8.0"
-pyth-solana-receiver-sdk = "0.3.0"
+# pyth-sdk-solana = "0.8.0"
+# pyth-solana-receiver-sdk = "0.3.0"
 solana-program = "=1.16.20"
 solend-sdk = { path = "../sdk" }
+oracles = { path = "../oracles" }
 spl-token = { version = "3.3.0", features=["no-entrypoint"] }
 static_assertions = "1.1.0"
-switchboard-on-demand = "0.1.12"
-switchboard-program = "0.2.0"
-switchboard-v2 = "0.1.3"
 
 [dev-dependencies]
 anchor-lang = "0.28.0"
@@ -37,6 +35,11 @@ serde_yaml = "0.8"
 thiserror = "1.0"
 bincode = "1.3.3"
 borsh = "0.10.3"
+pyth-sdk-solana = "0.8.0"
+pyth-solana-receiver-sdk = "0.3.0"
+switchboard-on-demand = "0.1.12"
+switchboard-program = "0.2.0"
+switchboard-v2 = "0.1.3"
 
 [lib]
 crate-type = ["cdylib", "lib"]
