+++ token-lending/sdk/Cargo.toml
@@ -13,15 +13,11 @@ borsh = "0.10.3"
 bytemuck = "1.5.1"
 num-derive = "0.3"
 num-traits = "0.2"
-pyth-sdk-solana = "0.8.0"
-pyth-solana-receiver-sdk = "0.3.0"
 solana-program = ">=1.9"
 spl-token = { version = "3.2.0", features=["no-entrypoint"] }
 static_assertions = "1.1.0"
-switchboard-on-demand = "0.1.8"
 thiserror = "1.0"
 uint = "=0.9.1"
-anchor-lang = "0.28.0"
 
 [dev-dependencies]
 assert_matches = "1.5.0"
