+++ token-lending/sdk/src/lib.rs
@@ -5,7 +5,6 @@
 pub mod error;
 pub mod instruction;
 pub mod math;
-pub mod oracles;
 pub mod state;
 
 // Export current sdk types for downstream users building with a different sdk version
